requirements.txt
mcp>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.31.0
uvicorn>=0.24.0
fastapi>=0.104.0